requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
freeze==1.0.10
pytz==2019.3
six==1.14.0
sqlparse==0.3.1

